Partition forest scene into Chathran reborn
I sincerely hope there will never again be a reason to commit this many
new files simultaneously. But, anyways. A startover is a unique chance
to do things right, and so I tried.

I created a Manager class for six large components of Chathran: low
level engine communication (appl), rendering (render), user interface
and user input (ui), vegetation (veg), terrain (terrain) and sky (sky).
There will be a few more eventually, but for now there was no need for
additional complexity.

In old Chathran I had serious issues with communication between
different components. I have tried to avoid such issues in the future by
creating an Interface for each Manager, which contains a bit of
functionality that other parts of Chathran need. The interface usually
contains of a few functions at most, allowing proper encapsulation of
everything else and a small and elegant interface for communication
across components.

After a long day's work and over a dozen new files, Chathran does
exactly what it did at the previous commit (save for a few manual
adjustments to the parameters of the terrain and the forest for a better
visual experience), and that is how it should be. However, it should now be ready
for expansion.