Facilitate initialization of vegetation from scripts
I've been working towards incorporating Lua scripting in Chathran. At
this time, I think it is not desirable to do anything computationally
intensive with Lua, because it's significantly slower, less suitable for
large and complex programs, and because it would add unnecessary
fragmentation (as well as the need for Chathran programmers to learn
Lua).

Instead, I decided upon an approach I'd like to abbreviate as SINE:
Script initialization, not evolution. The initialization of Chathran
should be done whenever possible via Lua scripts, which determine what
resources should be loaded, how these resources are composed, and where
in Chathran they are to be used. On the other hand, the evolution of
Chathran should be script-free except when *de novo* initialization is
required.

This implies a program structure that is nearly an engine: the game's
composition is described fully by the scripts used in its
initialization, and its evolution is carried out by applying the set of
rules that is generated by the scripts. Broadly speaking, this means
Chathran should contain carefully designed interfaces (for receiving
initialized resources resulting from Lua scripts) and a very general
method for time evolution which allows dozens of parameters to be set
(from the scripts) in order to allow for high complexity.

For today's commit, I worked on such an interface for tiled vegetation
(like the forest already present in earlier commits). I cannot link Lua
yet since I am presently unable to connect to Internet, but I prepared
functions in src/veg/drawpar.hpp that should enable me to receive a
function call in only Lua variables (strings, numbers, bools) and
convert it to all objects needed for the forest TiledHorde.