Link with Lua and add Lua scripts (PARTIAL)
For Chathran, we are going to assume using Lua 5.3 (the newest version as
of this commit) but probably earlier versions would work, too. For now
FindLua.cmake looks for version 5.3 only.

The Lua scripts I am also including in this commit are required for the
updated source code. However, to keep each commit clean, I postponed
committing the actual source code for using Lua until the last moment,
and not commit everything all at once.