Update LICENSE and add third party licenses
To avoid growing confusion about the terms under which Chathran is
available, and the licenses we should include to be correct, I compiled
a list of all libraries Chathran links to, and their respective
licenses. In particular, I added Selene's license, which is especially
important as I included its source code in the Chathran repository.