Implemented Lua scripting for loading vegetation (FINAL)
This is the final part of the commit series starting with the first
PARTIAL. In the following I will detail the intentions behind this and
the usage of Lua scripting in Chathran.

CORE ADDITIONS

The code previously committed in src/veg/drawpar.hpp is now actively
used (though renamed as src/veg/drawtiled.hpp). After all Managers have
initialized successfully, a Lua state is created by the Game class and
subsequently all C++ functions that need to be called by Lua are
registered by each Manager.

Then, Lua scripts recursively call each other, and initialization of
resources takes place through Lua calls on previously registered C++
functions. After all resources are initialized, Lua returns control to
C++ (after which Lua in principle won't be used anymore, at least until
we introduce deferred initialization) and the Game class calls some
finalizing functions for resources that are composed of several separate
Lua models (presently, Game must call vegManager.addToTiledHorde() to
combine all loaded LODs into a TiledHorde).

Lua-loaded resources are stored in very general template classes, e.g.
the whole vegetation category has just one structure for storing all
TiledHorde's it may have. Although placement algorithms need
improvement, this means the code for such vegetation is essentially
ready to handle very diverse vegetation consisting of many distinct
models.

REASONING

As remarked in the first commit of a few days ago, which didn't include
any actually used code since I couldn't link with Lua at the time,
Chathran will employ scripting as its primary (and basically only)
method of initializing. The current state shows how this can be applied
to loading of Vegetation (i.e. the forest we already had).

Thus far, loading of Forest meshes was done by the VegManager using
hard-coded locations of the files for defining the mesh, the diffuse
texture, the normal texture, etcetera. Furthermore, parameters such as
the size of a tile of the TiledHorde and the maximum visibility radius
were also hardcoded.

None of all these things should require compilation. They are simply
assets to Chathran and not integral to the execution and evolution of
Chathran. These things, which are set once and then remain unchanged
during Chathran execution should be independent of the code, and anyone
ought to be able to simply add content to Chathran without having to
touch a compiler.

The simplicity of such an approach is highlighted by the Lua scripts
included in the previous commit. One can copy-paste a Lua script, change
a handful of parameters, and include the name of the new script in the
load.lua in the same directory, after which Chathran will pick it up.

FINAL REMARKS

The primary remaining issues with Chathran's present state are an
unrealistic terrain, nonexistent interactions for vegetation placement
(to avoid vegetation being placed very close to each other), and the
absence of any UI, AI, animations, collision detection, etcetera. So a
lot of work remains to be done. But the way in which vegetation models
are now loaded, and the way their rendering is organized, is more
elegant than it has ever been in old Chathran.