Simplify Veg code by removing RGB/RGBA distinction
The Vegetation code distinguished between RGBTexture2D and
RGBATexture2D, while in practice only pointers were used. Since both
classes are eventually derived from the TextureInterface, it seems
acceptable to use a TextureInterface* to both and remove the
distinction, putting them in the same array and removing a template
parameter of the VegetationLodComponent.

Note that TextureInterface's destructor is virtual, and the only reason
for us to keep these pointers at all is for being able to free the
memory when the tiledHorde is freed.