Link Chathran against separate data directory
In this commit I 'link' (I use the term loosely, I admit) Chathran
against the data directory that contains the description of the game
(through Lua scripts, models, textures, and so on). I use CMake to set
the proper include directory and help Chathran find its data. Though
there is nothing to 'make' really (after all none of the contents of the
data repository are compiled, not even Lua, until runtime), this seems
like a neat method to find and include it.

The whole data mess is gone now and the repository should be
considerably cleaner. I do not anticipate a lot of work to be done on
Chathran itself in the coming weeks (with most of my efforts going into
decent terrains from Strata first). But I wanted to have this split
finished before any more work is done, and before I am integrating
Strata.

On that matter, I do not intend to split Strata's data directory into a
distinct repository since it will remain very simple and lightweight,
and only for the purpose of displaying the terrains generated.